+++ CHANGELOG.md
@@ -20,6 +20,10 @@
 - Added `\Wizaplace\SDK\AuthLog\AuthLogService::get`
 - Added `\Wizaplace\SDK\AuthLog\AuthLogService::search`
 
+### Bugfixes
+
+- Fixed missing Order status in `Wizaplace\SDK\Vendor\Order\OrderStatus` and `Wizaplace\SDK\Order\OrderStatus` 
+
 ## 1.52.0
 
 ### New features