+++ CHANGELOG.md
@@ -9,7 +9,33 @@
 - Added `\Wizaplace\SDK\Vendor\Order\OrderListFilter::byLastStatusChangeIsBefore`
 - Updated `\Wizaplace\SDK\Vendor\Order\OrderService::listOrders`
 
+## 1.57.0
+
+### Bugfixes
+
+- Added `Iban` and `Bic` to data sent by `Wizaplace\SDK\Company\CompanyService::unauthenticatedRegister()`
+- Fixed `\Wizaplace\SDK\Company\Company::__construct` with NAF Code `null` in constructor
+
+## 1.56.0
+
+### New features
+
+- Updated `\Wizaplace\SDK\Discussion\DiscussionService::submitContactRequest` add 3 optional params `$recipientEmail`, `$attachmentsUrls` and `$files`
+
+## 1.55.1
+
+### New features
+
+- Added `\Wizaplace\SDK\AuthLog\AuthLogService::get`
+- Added `\Wizaplace\SDK\AuthLog\AuthLogService::search`
+
+### Bugfixes
+
+- Fixed missing Order status in `Wizaplace\SDK\Vendor\Order\OrderStatus` and `Wizaplace\SDK\Order\OrderStatus` 
+
 ## 1.52.0
+
+### New features
 
 - Fixed `\Wizaplace\SDK\Catalog\Product` shippings into declination
 - Added `\Wizaplace\SDK\Company\AbstractCompanyRegistration::getNafCode`
@@ -104,7 +130,7 @@
 
 ## 1.44.13
 
- ### New features
+### New features
 
 - Added `\Wizaplace\SDK\Company\CompanyRegistration::getIban`
 - Added `\Wizaplace\SDK\Company\CompanyRegistration::setIban`
@@ -126,7 +152,7 @@
 ### Bugfixes
 
 - Fixed `\Wizaplace\SDK\Catalog\ProductOffer` with null `$data['divisions']`
-- Fixed `\Wizaplace\SDK\Catalog\ProductSummary::getOffers` return null 
+- Fixed `\Wizaplace\SDK\Catalog\ProductSummary::getOffers` return null
 
 ## 1.44.10